tests/wpt/web-platform-tests/ambient-light/AmbientLightSensor_browsing_context.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: Sensor readings must only be available in the top-level browsing context</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="support-iframe.html" id="frame" style="display:none">
</iframe>
<script>

async_test(function (t) {
  document.getElementById('frame').onload = t.step_func_done(function(event) {
    var iframe = document.getElementById('frame').contentWindow;
    let reading = iframe.document.getElementById('reading').value;
    assert_equals(reading, "null");
  });
}, "sensor readings can not be fired within iframes");

test(function() {
  let sensor = new AmbientLightSensor();
  sensor.start();
  var win = window.open('', '_blank');
  let reading = String(sensor.reading);
  win.close();
  sensor.stop();
  assert_equals(reading, "null");
}, "sensor readings can not be fired on the background tab");

</script>
